FOUND variable already set with proper casing.
This line was introduced as part of 2a1ed4b and was probably added for
reasons related to the issue resolved by 460faf8. However, the result
of merging these two PRs is that after successfully finding TinyXML2 and
setting both TinyXML2_FOUND and TINYXML2_FOUND (per the documentation in
find_package_handle_standard_args) we are then re-setting TinyXML2_FOUND
to the value of a now-unused variable: tinyxml2_FOUND resulting in it
being unset/empty again after being found.

The all-lower-case tinyxml2_FOUND / find_package(tinyxml2) is not used
within the ROS 2 sources (nor should it be).